+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## master (unreleased)
+## 0.3.0 (2020-11-22)
 
 ### New features